2999: CE.java (#869)
